requirements.txt
humanize==4.7.0
icoscp==0.1.18
netCDF4==1.6.4
pyyaml==6.0
requests==2.31.0
xarray==2023.6.0
icoscp_core==0.3.0

pip~=22.0.2
cryptography~=41.0.2
Jinja2~=3.1.2
keyring~=24.2.0
numpy~=1.25.0
pytz~=2023.3
setuptools~=59.6.0
cffi~=1.15.1
mypy~=1.7.0
packaging~=23.1
pandas~=2.0.3
MarkupSafe~=2.1.3
python-dateutil~=2.8.2
cftime~=1.6.2
six~=1.16.0